% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVU-RSBN.R
\name{corrNVU_VOR}
\alias{corrNVU_VOR}
\title{Calculate NVU orthodromic coordinates using VOR/DME}
\usage{
corrNVU_VOR(plan, leg, VOR, radialTo, dist, S = NA, cols = c("fix",
  "fixLat", "fixLon", "ZPY"))
}
\arguments{
\item{plan}{A flight plan, as a data.frame. Plans created with \code{\link{planNVU}(..., merge=TRUE)}
are accepted as input. At a minimum, the data frame.should have four columns:
waypoint names, latitudes, longitudes and ZPY (NVU course). The latter is needed to indirectly
identify the NVU legs, which may not correspond 1:1 to the legs of the flight plan (see
\code{points=} in \code{\link{planNVU}}).}

\item{leg}{The current leg of the flight plan, whose NVU parameters are to be corrected.
It is specified either as a number or as the initial fix.}

\item{VOR}{ID of the VOR to use.If multiple stations match, the correct one is
determined based on proximity to the initial fix of the leg.}

\item{radialTo}{Current \emph{inbound} radial (as indicated on the UShDB instrument of the Tu-154).}

\item{dist}{Distance in km, as measured by the DME}

\item{S}{Along track distance S, as currently indicated in the NVU computer. If supplied,
it must always be negative (as is the NVU convention).}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}
}
\value{
Returns the current orthodromic coordinates Z (cross track error) and S (along track
distance) as a named vector.These can be fed into the NVU computer to correct the course of the
aircraft.

Because S is changing rapidly as the aircraft travels, for convenience if S is supplied as an
argument, an extra element dS is added to the return vector indicating the difference between
actual (corrected) S minus indicated S. Add dS to the current S indication of the NVU to correct it.
}
\description{
This function calculates the current NVU orthodromic coordinates
(Z and S) based on radial and distance from a VOR/DME beacon.
}

