% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVU-RSBN.R
\name{corrRSBN}
\alias{corrRSBN}
\title{Calculate RSBN course correction}
\usage{
corrRSBN(lat1, lon1, lat2, lon2, latB, lonB)
}
\arguments{
\item{lat1, lon1}{Latitude and longitude of the start point}

\item{lat2, lon2}{Latitude and longitude of the end point}

\item{latB, lonB}{Latitude and longitude of the beacon}
}
\value{
A numeric vector of length 3, containing the values for Zm, Sm and YK.
}
\description{
This function returns the NVU course correction parameters (Zm, Sm, YK) given the latitude and longitude of
the start point, the end point and the RSBN beacon.
}
\details{
This function is used by \code{corrPlanRSBN()}.
}

