% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix2fix.R
\name{courseDiff}
\alias{courseDiff}
\title{Determine difference between two course headings}
\usage{
courseDiff(a, b)
}
\arguments{
\item{a, b}{Course headings, in degrees}
}
\value{
The difference a - b, in degrees
}
\description{
This function returs the difference (in degrees) between two course headings.
A negative sign means that a is to the left of b, a positive sign means that a is to the right of b.
Obviously the two course headings should be on the same reference, e.g. both magnetic or both true.
}
