% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgfsCom.R
\name{fgfsConnect}
\alias{fgfsConnect}
\title{Connect to FlightGear via telnet}
\usage{
fgfsConnect(host = "localhost", port = 5501)
}
\arguments{
\item{host}{Hostname/IP address of the computer running FlightGear and
the props interface (default: localhost).}

\item{port}{Port that the props server is listening on (default: 5501).}
}
\value{
A (socket) connection object, used with other package functions to
control the simulator. It is your responsibility to \code{\link{close}} it
when you don't need it anymore.
}
\description{
This function creates a socket connection to the FlightGear props (telnet)
interface, via which the property tree can be queried and/or modified.
}
\details{
As soon as the connection is opened, the interface is switched
to bulk data mode, and a help commend is issued. If FlightGear does not
respond within 5 seconds with the expected help message, the function
quits with an error. Otherwise the connection object is returned.
}
