% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgfsCom.R
\name{fgfsGet}
\alias{fgfsGet}
\title{Get (read) FlightGear property via telnet}
\usage{
fgfsGet(con, var, wait = 2)
}
\arguments{
\item{con}{A socket connection object created with \code{\link{fgfsConnect}}}

\item{var}{The property to read.}

\item{wait}{Maximum time waiting for a server response.}
}
\value{
The value of the property. This is always a string. If the
property does not exist on the tree, an empty string is returned. If the
server does not respond, a character vector of length zero is returned.
}
\description{
This function reads a property from the FlightGear property tree, via
the props (telnet) interface
}
