% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgfsCom.R
\name{fgfsSet}
\alias{fgfsSet}
\title{Set FlightGear property via telnet}
\usage{
fgfsSet(con, var, val)
}
\arguments{
\item{con}{A socket connection object created with \code{\link{fgfsConnect}}}

\item{var}{The name of the property.}

\item{val}{The value to set.}
}
\value{
None.
}
\description{
This function sets a FlightGear property via the props (telnet) interface.
}
\details{
This value does no error checking. If the property you try to
set does not exist on the tree, failure is silent.
}
