% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awys.R
\name{findAwy}
\alias{findAwy}
\title{Find airways in the database}
\usage{
findAwy(x)
}
\arguments{
\item{x}{A navaid/navfix as a character string.}
}
\value{
A character vector with the name(s) of the airway(s) that include the supplied navaid/navfix. A vector of length zero if no matching airway is found.
}
\description{
\code{findAwy} finds airways in the database that include the supplied navfix/navaid. Internally it uses \code{grep}.
}
