% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacons.R
\name{findBeacon}
\alias{findBeacon}
\title{Find beacon(s) in the database}
\usage{
findBeacon(x, type = c("VOR", "NDB", "TACAN", "RSBN"), exact.id = FALSE)
}
\arguments{
\item{x}{A character string containing a regular expression to be matched in the navaid ids or names.}

\item{type}{Type(s) of beacons to be returned. By default, return all possible beacons (VORs, NDBs, TACANs and RSBNs).}

\item{exact.id}{Match only on beacon id (as exact string).}
}
\value{
A data.frame of beacons found, with columns: \code{lat}, \code{lon}. \code{elev}, \code{freq}, \code{id}, \code{name}.
}
\description{
\code{findBeacon} finds navigational beacons in the database by searching for matches in the navaid id or name.
Internally it uses \code{grep}.
}
