% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fix2fix.R
\name{fix2fix}
\alias{fix2fix}
\title{Find the shortest route between two navfixes/navaids}
\usage{
fix2fix(startFix, endFix, limitArc = TRUE, narrowArc = FALSE)
}
\arguments{
\item{startFix}{Identifier for departure navfix/navaid.}

\item{endFix}{Identifier for destination navfix/navaid.}

\item{limitArc}{If TRUE, only search points at an angle of ±90 degrees towards the destination point in order to speed up the algorithm.}

\item{narrowArc}{If TRUE (and also limitArc==TRUE), search at a narrower angle of ±45 degrees towards the destination point.}
}
\value{
Returns a vector of identifiers corresponding to the determined route.
}
\description{
This function determines the shortest route between two points (navfixes/navaids)
}

