% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix2fix.R
\name{gcb}
\alias{gcb}
\title{Find the initial course for an orthodromic route}
\usage{
gcb(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude of departure point}

\item{lon1}{Longitude of departure point}

\item{lat2}{Latitude of destination point}

\item{lon2}{Longitude of destination point}
}
\value{
Returns the initial orthodromic course in degrees.
}
\description{
This function returs the initial course heading (in degrees) for an orthodromic course between two points.
}

