% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apts.R
\name{getAptFix}
\alias{getAptFix}
\title{Get the navaid or navfix nearest to an airport}
\usage{
getAptFix(airport)
}
\arguments{
\item{airport}{A 4-letter airport ICAO code (character vector of length 1).}
}
\value{
A string identifying the nearest navfix/navaid. Because identifiers are not globally unique, the string includes a number, which is the difference between latitude and longitude.
}
\description{
\code{getAptFix} finds the navaid or navfix that is closest to a given airport.
}
\seealso{
\code{\link{planRoute}}
}
