% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awys.R
\name{getAwy}
\alias{getAwy}
\title{Get airways data from the database}
\usage{
getAwy(x, fix = NA, exact = TRUE)
}
\arguments{
\item{x}{An airway name (as a character string), or part thereof.}

\item{fix}{An optional navaid/navfix name. If given, only the airways that include this are returned.}

\item{exact}{If FALSE, x is matched partially (i.e. all airways that contain x in their name are returned. If TRUE, x is matched exactly (i.e. only airways that are exactly named x will be returned). Defaults to TRUE.}
}
\value{
A list of data.frames. Each data.frame describes the segments of one airway, and its columns are: "fix1", "fix1Lat", "fix1Lon", "fix2", "fix2Lat", "fix2Lon", "isHigh", "baseFL", "topFL", "awy", "id1", "id2", "dist", "angle"
}
\description{
\code{getAwy} gets detailed airways info from the database. Internally it uses \code{grep}.
}
