% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/apts.R
\name{is.apt}
\alias{is.apt}
\title{Is the argument a valid airport ICAO code?}
\usage{
is.apt(x)
}
\arguments{
\item{x}{A character vector (possibly of length >1) of airport ICAO codes.}
}
\value{
Returns TRUE for each element of x that is found in the database of airport ICAO codes.
}
\description{
Checks if the supplied argument is a valid 4-letter airport ICAO code.
}
\seealso{
\code{\link{findApt}}, \code{\link{is.awy}}
}

