% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag.R
\name{magvar_apt}
\alias{magvar_apt}
\title{Calculate magnetic declination at a given airport}
\usage{
magvar_apt(x, date = Sys.Date())
}
\arguments{
\item{x}{A 4-letter ICAO airport code}

\item{date}{Date (as a Date object). Defaults to current date.}
}
\value{
Returns the magnetic declination (in degrees) at the specified airport.
}
\description{
Calculates magnetic declination at a given airport
}
\details{
This function calls \code{\link{magvar}}, which uses the NIMA WMM2005 (World Magnetic Model). This is the exact same algorithm that FlightGear (via the SimGear library) uses, and in fact \code{magvar} is a port of the corresponding SimGear C++ function: \url{https://github.com/FlightGear/simgear/blob/next/simgear/magvar/coremag.cxx}
}
