% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKml.R
\name{makeKml}
\alias{makeKml}
\title{Write a flight plan to a kml file}
\usage{
makeKml(fltplans, file, overwrite = TRUE)
}
\arguments{
\item{A}{flight plan data.frame, as returned from \code{\link{planRoute}}, or a list of such flight plans.}
}
\description{
This function writes a flight plan to a kml file, suitable for opening and viewing with Google Earth (or other kml viewer).
}
