% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRoute.R
\name{updateRoute}
\alias{updateRoute}
\title{Update (fill-in) a flight plan.}
\usage{
updateRoute(plan, cols = c("fix", "fixLat", "fixLon"))
}
\arguments{
\item{plan}{A flight plan, as a character vector (of fix names) or as a data.frame. Plans created with
\code{\link{planRoute}} are accepted as input. If a data.frame, it should have a minimum of three columns:
waypoint names, latitudes and longitudes. However, it is usually OK to have missing values (see details).}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}
}
\value{
Returns a data.frame containing the full flight plan, with the same columns as the one
returned by \code{\link{planRoute}}.
}
\description{
This function takes a data.frame of fixes and their coordinates and fills in the flight plan details
(courses, distances, airways, etc as in \code{\link{planRoute}}). This can be used to modify by hand
a plan created with \code{\link{planRoute}}, or even to manually create a plan from scratch (see examples).
}
\details{
Provided the fix names are correct and exist in the database, it is not necessary to provide
their coordinates; thus \code{plan} can be a character vector of fix names, or a data.frame with all
\code{NA}s on the coordinate columns. Even for non-unique fix names, the correct one will be chosen
based on proximity to the nearest fixes. Conversely, if you provide coordinates (in a data.frame),
then fixes don't have to be named. However if fixes are not named, then determining the
ICAO airway between them is not possible.

If you fly the Tu-154B, the flight plan created with \code{updateRoute} can be used with
\code{\link{planNVU}}, but if you have unnamed fixes you must use row numbers for the \code{points=} argument.
}

