% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn.R
\name{vnCorrNVU_VOR}
\alias{vnCorrNVU_VOR}
\title{"Virtual Navigator": Correct NVU orthodromic coordinates using VOR/DME}
\usage{
vnCorrNVU_VOR(con, plan, leg, VOR, UShDB = 2, DME = UShDB, apply = TRUE,
  cols = c("fix", "fixLat", "fixLon", "ZPY"))
}
\arguments{
\item{con}{A socket connection object created with \code{\link{fgfsConnect}}}

\item{plan}{A flight plan, as a data.frame. Plans created with \code{\link{planRoute}}
are accepted as input. At a minimum, the data frame.should have three columns:
waypoint names, orthodromic course headings (ZPY) and leg distances (S).}

\item{leg}{The next (i.e. upcoming) leg of the flight plan. Specified as waypoint name.}

\item{VOR}{The 3-letter identifier of the VOR beacon.}

\item{UShDB}{Which of the two UShDB needles shows the current VOR radial? Either 1 or 2.}

\item{DME}{Which DME radio to use? Either 1 or 2.}

\item{apply}{Tune the active NVU device to the calculated orthodromic coordinates? If \code{FALSE},
these are returned without adjusting the NVU device.}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}
}
\description{
Through the FlightGear props (telnet) interface and on the Tu-154B, obtain
radial and distance from a certain VOR/DME beacon, calculate the current
orthodromic coordinates S and Z, and (optionally) tune them into the active
NVU device.
}

