% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn.R
\name{vnNVUautoload}
\alias{vnNVUautoload}
\title{"Virtual Navigator": Autoload next flight plan leg to the inactive NVU device of the Tu-154B}
\usage{
vnNVUautoload(con, plan, leg, init = FALSE, poll = 10, cols = c("fix",
  "ZPY", "S"))
}
\arguments{
\item{con}{A socket connection object created with \code{\link{fgfsConnect}}}

\item{plan}{A flight plan, as a data.frame. Plans created with \code{\link{planRoute}}
are accepted as input. At a minimum, the data frame.should have three columns:
waypoint names, orthodromic course headings (ZPY) and leg distances (S).}

\item{leg}{The current leg of the flight plan. Specified as waypoint name.}

\item{init}{If \code{TRUE}, also tune the active NVU device to the start of the current leg.
You will only want to use this on the ground, before you takeoff.}

\item{poll}{Time (in seconds) between successive checks of whether a waypoint change happened.
In between, the function \code{\link{Sys.sleep}}s to conserve CPU cycles.}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}
}
\value{
None
}
\description{
Through the FlightGear props (telnet) interface and on the Tu-154B, this function
monitors the NVU and with every waypoint change, it loads the next
flight plan leg to the inactive NVU device.
}
\details{
This function loads each leg of your journey to the NVU, as your navigator would do,
so you can focus on other things. It stops if there's no other leg to be loaded, or if you press
Ctrl-C. For example you can stop it to correct your position with \code{\link{vnCorrNVU_VOR}},
and then run it again from the current leg.
}
