% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn.R
\name{vnNVUloadNextLeg}
\alias{vnNVUloadNextLeg}
\title{"Virtual Navigator": Tune the Tu-154B NVU to the next leg of a flight plan}
\usage{
vnNVUloadNextLeg(con, plan, leg, cols = c("fix", "ZPY", "S"))
}
\arguments{
\item{con}{A socket connection object created with \code{\link{fgfsConnect}}}

\item{plan}{A flight plan, as a data.frame. Plans created with \code{\link{planRoute}}
are accepted as input. At a minimum, the data frame.should have three columns:
waypoint names, orthodromic course headings (ZPY) and leg distances (S).}

\item{leg}{The next (i.e. upcoming) leg of the flight plan. Specified as waypoint name.
If ZPY and S for that leg are \code{NA}, nothing happens.}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}
}
\description{
Through the FlightGear props (telnet) interface, load the NVU parameters for the
next (i.e. upcoming) leg of a flight plan to the inactive NVU device of the Tu-154B.
}
