% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn.R
\name{vnNVUloadRSBN}
\alias{vnNVUloadRSBN}
\title{"Virtual Navigator": Load RSBN correction parameters on a Tu-154B}
\usage{
vnNVUloadRSBN(con, legs, row = 1, cols = c("Zm", "Sm", "YK"))
}
\arguments{
\item{con}{A socket connection object created with \code{\link{fgfsConnect}}}

\item{legs}{A data.frame of leg-correction pairs such as those created by
\code{\link{corrPlanRSBN}}. At a minimum, the data frame.should have three columns:
cross-track distance (Zm), along-track distance (Sm) and map angle (YK).}

\item{row}{Which row of the data.frame to use?}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}
}
\description{
Through the FlightGear props (telnet) interface, tune RSBN correction parameters
on the appropriate devices (B-52 panel, B-8M map angle instrument) of the Tu-154B,
taking them from a data.frame of leg-correction pairs (of the type created by
\code{\link{corrPlanRSBN}}.
}

