% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn.R
\name{vnNVUtuneRSBN}
\alias{vnNVUtuneRSBN}
\title{"Virtual Navigator": Tune RSBN correction parameters on the Tu-154}
\usage{
vnNVUtuneRSBN(con, Zm = NA, Sm = NA, YK = NA)
}
\arguments{
\item{con}{A socket connection object created with \code{\link{fgfsConnect}}}

\item{Zm}{Cross-track distance (Zm), in km. If \code{NA}, the already tuned Zm is not modified.}

\item{Sm}{Along-track distance (Sm), in km. If \code{NA}, the already tuned Sm is not modified.}

\item{YK}{Map angle (YK), in degrees. This is rounded to the first decimal when
passing to FlightGear. If \code{NA}, the already tuned YK is not modified.}
}
\description{
Through the FlightGear props (telnet) interface, tune RSBN correction parameters
on the appropriate devices (B-52 panel, B-8M map angle instrument) of the Tu-154.
}

