% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix2fix.R
\name{findFixes}
\alias{findFixes}
\title{Find navigational fix(es) by name}
\usage{
findFixes(x, refPoint = NULL, type = c("airport", "VOR", "TACAN", "NDB",
  "RSBN"))
}
\arguments{
\item{x}{Fix id to search for.}

\item{refPoint}{Optional reference point. A list containing elements \code{lat} and \code{lon},
holding the latitude and longitude in degrees.}
}
\value{
A data.frame (with zero rows if no fix is found) with columns \code{lat}, \code{lon}, \code{elev},
\code{freq}, \code{range}, \code{id}, \code{name}, \code{type}.
}
\description{
This function finds navigational fixes (fixes or navaids) by name (id). If there are many fixes
with the same name, they are all returned unless an optional reference point is supplied, in which
case only the fix nearest to the reference point is returned.
}

