% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVU-RSBN.R
\name{planNVU}
\alias{planNVU}
\title{Obtain NVU navigational data for a flight plan}
\usage{
planNVU(plan, points = NULL, cols = c("fix", "fixLat", "fixLon"),
  init.mag = FALSE, merge = FALSE)
}
\arguments{
\item{plan}{A flight plan, as a data.frame. Plans created with \code{\link{planRoute}}
are accepted as input. At a minimum, the data frame.should have three columns:
waypoint names, latitudes and longitudes.}

\item{points}{An optional vector of row numbers or waypoint names to select from the flight plan.}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}

\item{init.mag}{At the initial waypoint, does the HSI show magnetic heading (instead of true heading)?}

\item{merge}{Merge the NVU data into \code{plan}?}
}
\value{
If \code{merge = FALSE}, returns a matrix with columns \code{ZPY}, \code{S} and \code{da}.
\code{ZPY} (orthodromic HSI course heading) and \code{S} (course distance, negated) can be fed
to the NVU panels. \code{da} is the azimuth correction at each waypoint.

If \code{merge = TRUE} the above columns are added to the data.frame \code{plan}.
}
\description{
This function gets a flight plan as input (i.e. a list of waypoints and
geographical coordinates) and calculates appropriate data to feed the
NVU navigational computer of the Tu-154B.
}

