% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/planRoute.R
\name{planRoute}
\alias{planRoute}
\title{Create a flight plan route between two airports.}
\usage{
planRoute(aptFrom, aptTo, fixFrom = NA, fixTo = NA, fixes = character(),
  limitArc = TRUE, narrowArc = FALSE)
}
\arguments{
\item{aptFrom}{Departure airport (4-letter ICAO code).}

\item{aptTo}{Destination airport (4-letter ICAO code).}

\item{fixFrom}{Initial departure navfix (after SID procedure). If not supplied manually, the navfix nearest to the departure airport will be chosen automatically, using \code{\link{getAptFix}}.}

\item{fixTo}{Final destination navfix (before STAR procedure). If not supplied manually, the navfix nearest to the destination airport will be chosen automatically, using \code{\link{getAptFix}}}

\item{fixes}{Optional vector of intermediate navfixes. Use these to speed up the algorithm, or to manually adjust the desired route.}

\item{limitArc}{If TRUE, only search points at an angle of ±90 degrees towards the destination point in order to speed up the algorithm.}

\item{narrowArc}{If TRUE (and also limitArc==TRUE), search at a narrower angle of ±45 degrees towards the destination point.}
}
\value{
Returns a flight plan stored on a data.frame with the following columns:
\describe{
 \item{fix}{The name each point (airport/navaid/navfix) in the flight plan.}
 \item{freq}{Radio frequency (for navaids only).}
 \item{range}{Reception range (for navaids only).}
 \item{fullName}{Full name }
 \item{fixLat}{Latitude of the point.}
 \item{fixLon}{Longitude of the point.}
 \item{elevation}{Elevation (for navaids or airports).}
 \item{awy}{Airway identifier for the following segment (up to the next point in the plan), or SID/STAR.}
 \item{dist}{Distance (in km) (up to the next point in the plan).}
 \item{baseFL}{Base flight level for the following segment (up to the next point in the plan).}
 \item{topFL}{Top flight level for the following segment (up to the next point in the plan).}
 \item{course}{Course (true) towards the next point.}
 \item{altawy}{Alternative airway for the following segment, along with its permitted flight level range.}
}
}
\description{
This function determines the shortest route between two points (navfixes/navaids)
}

