% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRoute.R
\name{updateRoute}
\alias{updateRoute}
\title{Update (fill-in) a flight plan.}
\usage{
updateRoute(plan, cols = c("fix", "fixLat", "fixLon"))
}
\arguments{
\item{plan}{A flight plan, as a character vector (of fix names) or as a data.frame. Plans created with
\code{\link{planRoute}} are accepted as input. If a data.frame, it should have a minimum of three columns:
waypoint names, latitudes and longitudes. However, it is usually OK to have missing values (see details).}

\item{cols}{An optional vector of column names, if these are named differently in \code{plan}}
}
\value{
Returns a data.frame containing the full flight plan, with the same columns as the one
returned by \code{\link{planRoute}}.
}
\description{
This function takes a data.frame of fixes and their coordinates and fills in the flight plan details
(courses, distances, airways, etc as in \code{\link{planRoute}}). This can be used to modify by hand
a plan created with \code{\link{planRoute}}, or even to manually create a plan from scratch (see examples).
}
\details{
Provided the fix names are correct and exist in the database, it is not necessary to provide
their coordinates; thus \code{plan} can be a character vector of fix names, or a data.frame with all
\code{NA}s on the coordinate columns. Even for non-unique fix names, the correct one will be chosen
based on proximity to the nearest fixes. Conversely, if you provide coordinates (in a data.frame),
then fixes don't have to be named. However if fixes are not named, then determining the
ICAO airway between them is not possible.

If you fly the Tu-154B, the flight plan created with \code{updateRoute} can be used with
\code{\link{planNVU}}, but if you have unnamed fixes you must use row numbers for the \code{points=} argument.
}
\examples{
# Fly from Athens to Ioannina, manually setting the waypoints
plan <- updateRoute(c("LGAV","KOR","PIKAD","IXONI","GARTA","LGIO"))


# Create a flight plan from Moscow (Domodedovo) to Norilsk (Alykel)
plan0 <- planRoute("UUDD", "UOOO", "WT", "BI01")

# We will manually add waypoints for the final appropach
# (taken from the approach plate):

# Take only the first three columns, and add space before the final row
plan1 <- plan0[c(1:(nrow(plan0)-1), rep(NA,3), nrow(plan0)), 1:3]

# First waypoint is Norilsk NDB (535 BF):
bf <- findFixes("BF", refPoint=as.list(findApt("UOOO")[,c("lat","lon")]), type="NDB")
plan1[nrow(plan1)-4,] <- bf[,c("id","lat","lon")]

# Add coordinates manually for the remaining points
plan1[nrow(plan1)-3,] <- list("D10.0 NOR", 69 + 23.2/60, 87 + 42.5/60)
plan1[nrow(plan1)-2,] <- list("D10.8 NOR", 69 + 26.5/60, 87 + 37.5/60)
plan1[nrow(plan1)-1,] <- list(NA, 69 + 24.1/60, 87 + 30/60)

# Now fill in the rest of the flight plan:
plan2 <- updateRoute(plan1)
print(plan2)

}
